Remove or comment out unneeded postgres/database code
The plan for now is to use SQL Server on my dev machine and then use
SQL Server in an AWS RDS instance for production. So the postgres code
is no longer needed. I also want to simplify some of the database code,
so I commented out some migration code from the `docker-entrypoint.sh`
file that I'm not planning on using for now.